+++ client/src/lib/state/index.js
@@ -1,7 +1,5 @@
 const state = require("./state");
-const { storageChangeDisabledCallback } = require("./utils");
 
 module.exports = {
   state,
-  storageChangeDisabledCallback,
 };