+++ unit-tests/forms/wizard.test.js
@@ -57,7 +57,7 @@ describe("setup wizard", () => {
         enable_power_vs: false,
         enable_classic: false,
         power_vs_zones: [],
-        craig_version: "1.12.0",
+        craig_version: "1.12.1",
         power_vs_high_availability: false,
       },
       resource_groups: [
@@ -971,7 +971,7 @@ describe("setup wizard", () => {
         enable_power_vs: false,
         enable_classic: false,
         power_vs_zones: [],
-        craig_version: "1.12.0",
+        craig_version: "1.12.1",
         power_vs_high_availability: false,
       },
       resource_groups: [
@@ -1884,7 +1884,7 @@ describe("setup wizard", () => {
         enable_power_vs: false,
         enable_classic: false,
         power_vs_zones: [],
-        craig_version: "1.12.0",
+        craig_version: "1.12.1",
         power_vs_high_availability: false,
       },
       resource_groups: [
@@ -2391,7 +2391,7 @@ describe("setup wizard", () => {
         enable_power_vs: false,
         enable_classic: false,
         power_vs_zones: [],
-        craig_version: "1.12.0",
+        craig_version: "1.12.1",
         power_vs_high_availability: false,
       },
       resource_groups: [
@@ -2889,7 +2889,7 @@ describe("setup wizard", () => {
         enable_power_vs: false,
         enable_classic: false,
         power_vs_zones: [],
-        craig_version: "1.12.0",
+        craig_version: "1.12.1",
         power_vs_high_availability: false,
       },
       resource_groups: [
@@ -3464,7 +3464,7 @@ describe("setup wizard", () => {
         enable_power_vs: false,
         enable_classic: false,
         power_vs_zones: [],
-        craig_version: "1.12.0",
+        craig_version: "1.12.1",
       },
       resource_groups: [
         { use_prefix: true, name: "service-rg", use_data: false },
@@ -3630,7 +3630,7 @@ describe("setup wizard", () => {
         enable_power_vs: false,
         enable_classic: false,
         power_vs_zones: [],
-        craig_version: "1.12.0",
+        craig_version: "1.12.1",
         power_vs_high_availability: false,
       },
       resource_groups: [
@@ -4197,7 +4197,7 @@ describe("setup wizard", () => {
         enable_power_vs: false,
         enable_classic: false,
         power_vs_zones: [],
-        craig_version: "1.12.0",
+        craig_version: "1.12.1",
         power_vs_high_availability: false,
       },
       resource_groups: [
@@ -4738,7 +4738,7 @@ describe("setup wizard", () => {
         enable_power_vs: false,
         enable_classic: false,
         power_vs_zones: [],
-        craig_version: "1.12.0",
+        craig_version: "1.12.1",
         power_vs_high_availability: false,
       },
       resource_groups: [
@@ -5284,7 +5284,7 @@ describe("setup wizard", () => {
         enable_classic: false,
         enable_classic: false,
         power_vs_zones: [],
-        craig_version: "1.12.0",
+        craig_version: "1.12.1",
       },
       resource_groups: [
         { use_prefix: true, name: "service-rg", use_data: false },
@@ -5772,7 +5772,7 @@ describe("setup wizard", () => {
         enable_power_vs: true,
         enable_classic: false,
         power_vs_zones: ["dal10"],
-        craig_version: "1.12.0",
+        craig_version: "1.12.1",
         power_vs_high_availability: false,
       },
       resource_groups: [
@@ -6270,7 +6270,7 @@ describe("setup wizard", () => {
         enable_power_vs: true,
         enable_classic: false,
         power_vs_zones: ["dal10"],
-        craig_version: "1.12.0",
+        craig_version: "1.12.1",
         power_vs_high_availability: false,
       },
       resource_groups: [